Make additional directory-structure changes
No impactful code changes; just moving files around in the directory
structure.
