+++ src/App.tsx
@@ -7,20 +7,20 @@ import { RootState } from "./app/store/store";
 import { useAppDispatch, useAppSelector } from "./app/hooks/store";
 import { setApplicationTheme } from "./app/store/userSettingsSlice";
 import { useAppGlobals } from "./app/hooks/useAppGlobals";
-import RootLayout from "./components/layout/RootLayout";
-import AlbumsScreen from "./components/layout/screens/AlbumsScreen";
-import ArtistsScreen from "./components/layout/screens/ArtistsScreen";
-import FavoritesScreen from "./components/layout/screens/FavoritesScreen";
-import CurrentTrackScreen from "./components/layout/screens/CurrentTrackScreen";
-import PresetsScreen from "./components/layout/screens/PresetsScreen";
-import PlaylistScreen from "./components/layout/screens/PlaylistScreen";
-import PlayheadManager from "./components/managers/PlayheadManager";
-import MediaGroupsManager from "./components/managers/MediaGroupsManager";
-import MediaSourceManager from "./components/managers/MediaSourceManager";
-import TracksScreen from "./components/layout/screens/TracksScreen";
-import WebsocketManager from "./components/managers/WebsocketManager";
+import RootLayout from "./components/app/layout/RootLayout";
+import AlbumsScreen from "./components/features/AlbumsScreen";
+import ArtistsScreen from "./components/features/ArtistsScreen";
+import FavoritesScreen from "./components/features/FavoritesScreen";
+import CurrentTrackScreen from "./components/features/CurrentTrackScreen";
+import PresetsScreen from "./components/features/PresetsScreen";
+import PlaylistScreen from "./components/features/PlaylistScreen";
+import PlayheadManager from "./components/app/managers/PlayheadManager";
+import MediaGroupsManager from "./components/app/managers/MediaGroupsManager";
+import MediaSourceManager from "./components/app/managers/MediaSourceManager";
+import TracksScreen from "./components/features/TracksScreen";
+import WebsocketManager from "./components/app/managers/WebsocketManager";
 import ErrorBoundary from "./components/app/ErrorBoundary";
-import StatusScreen from "./components/layout/screens/StatusScreen";
+import StatusScreen from "./components/features/StatusScreen";
 
 export default function App() {
     const dispatch = useAppDispatch();
