+++ src/components/features/FavoritesScreen.tsx
@@ -1,17 +1,19 @@
 import React, { FC, useEffect } from "react";
 import { Box, Stack } from "@mantine/core";
 
-import { useAppGlobals } from "../../../app/hooks/useAppGlobals";
-import { useAppDispatch, useAppSelector } from "../../../app/hooks/store";
-import { RootState } from "../../../app/store/store";
-import { useWindowScroll } from "../../../app/hooks/useWindowScroll";
-import { setFavoritesScrollPosition } from "../../../app/store/internalSlice";
-import FavoritesControls from "../../features/favorites/FavoritesControls";
-import FavoritesWall from "../../features/favorites/FavoritesWall";
-import ScreenHeader from "../ScreenHeader";
+import { useAppGlobals } from "../../app/hooks/useAppGlobals";
+import { useAppDispatch, useAppSelector } from "../../app/hooks/store";
+import { RootState } from "../../app/store/store";
+import { useWindowScroll } from "../../app/hooks/useWindowScroll";
+import { setFavoritesScrollPosition } from "../../app/store/internalSlice";
+import FavoritesControls from "./favorites/FavoritesControls";
+import FavoritesWall from "./favorites/FavoritesWall";
+import ScreenHeader from "../app/layout/ScreenHeader";
 
 // ================================================================================================
 // Favorites screen top-level layout.
+//
+// Contains a <ScreenHeader>, <FavoritesControls>, and <FavoritesWall>.
 // ================================================================================================
 
 const FavoritesScreen: FC = () => {
