+++ src/components/features/PresetsScreen.tsx
@@ -1,17 +1,19 @@
 import React, { FC, useEffect } from "react";
 import { Box, Stack } from "@mantine/core";
 
-import { useAppGlobals } from "../../../app/hooks/useAppGlobals";
-import { useAppDispatch, useAppSelector } from "../../../app/hooks/store";
-import { RootState } from "../../../app/store/store";
-import { useWindowScroll } from "../../../app/hooks/useWindowScroll";
-import { setPresetsScrollPosition } from "../../../app/store/internalSlice";
-import PresetsControls from "../../features/presets/PresetsControls";
-import PresetsWall from "../../features/presets/PresetsWall";
-import ScreenHeader from "../ScreenHeader";
+import { useAppGlobals } from "../../app/hooks/useAppGlobals";
+import { useAppDispatch, useAppSelector } from "../../app/hooks/store";
+import { RootState } from "../../app/store/store";
+import { useWindowScroll } from "../../app/hooks/useWindowScroll";
+import { setPresetsScrollPosition } from "../../app/store/internalSlice";
+import PresetsControls from "./presets/PresetsControls";
+import PresetsWall from "./presets/PresetsWall";
+import ScreenHeader from "../app/layout/ScreenHeader";
 
 // ================================================================================================
 // Presets screen top-level layout.
+//
+// Contains a <ScreenHeader>, <PresetsControls>, and <PresetsWall>.
 // ================================================================================================
 
 const PresetsScreen: FC = () => {
