+++ src/components/shared/mediaDisplay/TrackLyricsModal.tsx
@@ -1,33 +1,29 @@
 import React, { FC } from "react";
-import { Modal, Stack } from "@mantine/core";
+import { Modal, ScrollArea, Stack } from "@mantine/core";
 
 import { Track } from "../../../app/types";
 import { useAppGlobals } from "../../../app/hooks/useAppGlobals";
-import MediaSummaryBanner from "../../shared/textDisplay/MediaSummaryBanner";
-import Waveform from "../currentTrack/Waveform";
+import MediaSummaryBanner from "../textDisplay/MediaSummaryBanner";
+import TrackLyrics from "./TrackLyrics";
 
 // ================================================================================================
-// Show a Track's waveform as a modal.
+// Show a Track's lyrics as a modal.
 // ================================================================================================
 
-type TrackWaveformModalProps = {
+type TrackLyricsModalProps = {
     track: Track;
     opened: boolean;
     onClose?: () => void;
 };
 
-const TrackWaveformModal: FC<TrackWaveformModalProps> = ({
-    track,
-    opened,
-    onClose = undefined,
-}) => {
+const TrackLyricsModal: FC<TrackLyricsModalProps> = ({ track, opened, onClose = undefined }) => {
     const { APP_MODAL_BLUR } = useAppGlobals();
 
     return (
         <Modal
             title={track.title}
             centered
-            size="xl"
+            size="lg"
             radius={7}
             overlayProps={{ blur: APP_MODAL_BLUR }}
             opened={opened}
@@ -35,10 +31,13 @@ const TrackWaveformModal: FC<TrackWaveformModalProps> = ({
         >
             <Stack>
                 <MediaSummaryBanner media={track} />
-                <Waveform trackId={track.id} width={2048} height={700} showProgress={false} />
+
+                <ScrollArea h={`calc(80vh - 275px)`}>
+                    <TrackLyrics trackId={track.id} />
+                </ScrollArea>
             </Stack>
         </Modal>
     );
 };
 
-export default TrackWaveformModal;
+export default TrackLyricsModal;
