+++ src/components/shared/playbackControls/CurrentMediaDetails.tsx
